requirements.txt
Flask
Flask-Login
Flask-Script
Werkzeug
pysqlite
Jinja2
SQLAlchemy
pyjwt
Flask-OAuthlib
